Initial addition of input blocks
