+++ examples/android/android-simple/app/src/main/AndroidManifest.xml
@@ -1,24 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!-- BEGIN_INCLUDE(manifest) -->
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-        package="com.ethicalml.komputejni">
+        package="com.ethicalml.kompute">
 
-<!--    <application android:label="@string/app_name"-->
-<!--      android:theme="@android:style/Theme.NoTitleBar.Fullscreen">-->
-<!--        &lt;!&ndash; Our activity is the built-in NativeActivity framework class.-->
-<!--             This will take care of integrating with our NDK code. &ndash;&gt;-->
-<!--        <activity android:name="android.app.NativeActivity"-->
-<!--                android:label="@string/app_name"-->
-<!--                android:configChanges="orientation|keyboardHidden">-->
-<!--            &lt;!&ndash; Tell NativeActivity the name of or .so &ndash;&gt;-->
-<!--            <meta-data android:name="android.app.lib_name"-->
-<!--                    android:value="kompute_android" />-->
-<!--            <intent-filter>-->
-<!--                <action android:name="android.intent.action.MAIN" />-->
-<!--                <category android:name="android.intent.category.LAUNCHER" />-->
-<!--            </intent-filter>-->
-<!--        </activity>-->
-<!--        </application>-->
         <application
             android:allowBackup="true"
             android:icon="@mipmap/ic_launcher"
