+++ examples/android/android-simple/app/src/main/res/mipmap-anydpi-v26/kompute_icon.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
+    <background android:drawable="@color/kompute_icon_background"/>
+    <foreground android:drawable="@mipmap/kompute_icon_foreground"/>
+</adaptive-icon>
