+++ examples/android/android-simple/app/src/main/res/values/kompute_icon_background.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <color name="kompute_icon_background">#FFFFFF</color>
+</resources>
